requirements.txt
streamlit
codecarbon
psutil
pandas
openai
requests
google-genai
groq
python-dotenv